Rename to .mjs
